+++ src/sync/all_forks.rs
@@ -1402,11 +1402,12 @@ impl<TBl, TRq, TSrc> AddBlockVacant<TBl, TRq, TSrc> {
                 None => break,
             };
 
-            self.inner
+            // TODO: restore this block of code; it is extremely complicated because it is unclear which source-block combinations we can add and keep without making memory usage explode
+            /*self.inner
                 .inner
                 .inner
                 .blocks
-                .remove_sources_known_block(height, &hash);
+                .remove_sources_known_block(height, &hash);*/
             self.inner
                 .inner
                 .inner
@@ -1631,10 +1632,11 @@ impl<'a, TBl, TRq, TSrc> AnnouncedBlockUnknown<'a, TBl, TRq, TSrc> {
                 None => break,
             };
 
-            self.inner
+            // TODO: restore this block of code; it is extremely complicated because it is unclear which source-block combinations we can add and keep without making memory usage explode
+            /*self.inner
                 .inner
                 .blocks
-                .remove_sources_known_block(height, &hash);
+                .remove_sources_known_block(height, &hash);*/
             self.inner
                 .inner
                 .blocks