+++ bin/wasm-node/javascript/src/instance/raw-instance.ts
@@ -39,7 +39,6 @@ export interface Config {
     onWasmPanic: (message: string) => void,
     logCallback: (level: number, target: string, message: string) => void,
     jsonRpcResponsesNonEmptyCallback: (chainId: number) => void,
-    databaseContentCallback: (data: string, chainId: number) => void,
     currentTaskCallback?: (taskName: string | null) => void,
     cpuRateLimit: number,
 }